+++ ChangeLog.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Better default HAProxy configuration
 - Update versions of k3s, metallb, and haproxy
 - Remove distro rpm repo and replace via ansible
 - Support HA k3s