+++ ChangeLog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.7.0] - 2023-10-04
 - Better default HAProxy configuration
 - Update versions of k3s, metallb, and haproxy
 - Remove distro rpm repo and replace via ansible